+++ src/pudl/analysis/allocate_gen_fuel.py
@@ -252,7 +252,6 @@ def allocate_gen_fuel_asset_factory(
         gens: pd.DataFrame,
     ) -> pd.DataFrame:
         """Allocate net gen from gen_fuel to generator/energy_source_code level."""
-        pd.options.mode.copy_on_write = True
         gf, bf, gen, bga, gens = select_input_data(
             gf=gf, bf=bf, gen=gen, bga=bga, gens=gens
         )
@@ -1841,7 +1840,7 @@ def distribute_annually_reported_data_to_months_if_annual(
                 )
             )
 
-        reporters = df.copy().pipe(assign_plant_year)
+        reporters = df.pipe(assign_plant_year)
         # get a count of the number of missing values in a year
         key_columns_annual = ["plant_year"] + [
             col for col in key_columns if col != "report_date"