+++ src/pudl/analysis/plant_parts_eia.py
@@ -115,7 +115,7 @@ combined-cycle unit and the gas-turbine.
 ... })
 >>> df_plant_part
     plant_id_eia           plant_part    unit_id_pudl    prime_mover_code    capacity_mw
-0              1           plant_unit               1                <NA>            200
+0              1           plant_unit               1                 NaN            200
 1              1    plant_prime_mover            <NA>                  GT            150
 
 In this case, the ``plant_unit`` record would have a null
@@ -958,8 +958,8 @@ class PlantPart:
         # id_cols = PLANT_PARTS[self.part_name]["id_cols"]
         # split up the 'owned' slices from the 'total' slices.
         # this is because the aggregations are different
-        part_own = gens_mega.loc[gens_mega.ownership_record_type == "owned"].copy()
-        part_tot = gens_mega.loc[gens_mega.ownership_record_type == "total"].copy()
+        part_own = gens_mega.loc[gens_mega.ownership_record_type == "owned"]
+        part_tot = gens_mega.loc[gens_mega.ownership_record_type == "total"]
         if len(gens_mega) != len(part_own) + len(part_tot):
             raise AssertionError(
                 "Error occurred in breaking apart ownership types."
@@ -1264,8 +1264,7 @@ class AddConsistentAttributes(AddAttribute):
             logger.debug(f"{attribute_col} already here.. ")
             return part_df
 
-        record_df = gens_mega.copy()
-        record_df = self.assign_col(record_df)
+        record_df = self.assign_col(gens_mega)
 
         consistent_records = self.get_consistent_qualifiers(record_df)
 
@@ -1338,7 +1337,7 @@ class AddPriorityAttribute(AddAttribute):
         gens_mega = self.assign_col(gens_mega)
         logger.debug(f"getting max {attribute_col}")
         consistent_records = pudl.helpers.dedupe_on_category(
-            gens_mega.copy()[self.base_cols + [attribute_col]],
+            gens_mega[self.base_cols + [attribute_col]],
             self.base_cols,
             attribute_col,
             PRIORITY_ATTRIBUTES_DICT[attribute_col],