+++ src/pudl/output/eia923.py
@@ -140,7 +140,7 @@ def drop_ytd_for_annual_tables(df: pd.DataFrame, freq: str) -> pd.DataFrame:
             "Removing rows where data_maturity is incremental_ytd to avoid "
             "aggregation errors."
         )
-        df = df.loc[df["data_maturity"] != "incremental_ytd"].copy()
+        df = df.loc[df["data_maturity"] != "incremental_ytd"]
     return df