+++ src/pudl/transform/classes.py
@@ -642,9 +642,8 @@ def nullify_outliers(col: pd.Series, params: ValidRange) -> pd.Series:
 
     The column is coerced to be numeric.
     """
-    # Surprisingly, pd.to_numeric() did *not* return a copy of the series!
-    col = col.copy()
-    col = pd.to_numeric(col, errors="coerce")
+    # pd.to_numeric() returns the same object, not a copy, when already numeric.
+    col = pd.to_numeric(col, errors="coerce").copy()
     col[~col.between(params.lower_bound, params.upper_bound)] = np.nan
     return col
 
@@ -899,7 +898,7 @@ def drop_invalid_rows(df: pd.DataFrame, params: InvalidRows) -> pd.DataFrame:
     invalids = cols_to_check.apply(_col_is_invalid)
     mask = ~(invalids.all(axis="columns"))
     # Mask the input dataframe and make a copy to avoid returning a slice.
-    df_out = df[mask].copy()
+    df_out = df[mask]
     logger.info(
         f"{1 - (len(df_out) / pre_drop_len):.1%} of records ({pre_drop_len - len(df_out)} "
         f"rows) contain only {params.invalid_values} values in required columns. "