+++ src/pudl/transform/eia860.py
@@ -35,8 +35,7 @@ def _core_eia860__ownership(raw_eia860__ownership: pd.DataFrame) -> pd.DataFrame
     """
     # Preliminary clean and get rid of unnecessary 'year' column
     own_df = (
-        raw_eia860__ownership.copy()
-        .pipe(pudl.helpers.standardize_na_values)
+        raw_eia860__ownership.pipe(pudl.helpers.standardize_na_values)
         .pipe(pudl.helpers.convert_to_date)
         .drop(columns=["year"])
     )
@@ -199,9 +198,6 @@ def _core_eia860__generators(
     Returns:
         Cleaned ``_core_eia860__generators`` dataframe ready for harvesting.
     """
-    # Groupby objects were creating chained assignment warning that is N/A
-    pd.options.mode.chained_assignment = None
-
     # There are three sets of generator data reported in the EIA860 table,
     # planned, existing, and retired generators. We're going to concatenate
     # them all together into a single big table, with a column that indicates
@@ -293,9 +289,8 @@ def _core_eia860__generators(
         .pipe(
             pudl.helpers.simplify_strings,
             columns=["rto_iso_lmp_node_id", "rto_iso_location_wholesale_reporting_id"],
-            copy=False,
         )
-        .pipe(pudl.helpers.convert_to_date, copy=False)
+        .pipe(pudl.helpers.convert_to_date)
     )
     # This manual fix is required before encoding because there's not a unique mapping
     # PA -> PACW in Oregon
@@ -308,7 +303,7 @@ def _core_eia860__generators(
         (gens_df.state == "UT") & (gens_df.balancing_authority_code_eia == "PA"),
         "balancing_authority_code_eia",
     ] = "PACE"
-    gens_df = PUDL_PACKAGE.encode(gens_df, copy=False)
+    gens_df = PUDL_PACKAGE.encode(gens_df)
 
     # spot fix one proposed fusion plant's technology description. It's assumed to be wrong
     # because we learned via pudl.output.eia.fill_generator_technology_description
@@ -673,7 +668,7 @@ def _core_eia860__plants(raw_eia860__plant: pd.DataFrame) -> pd.DataFrame:
         p_df.county.str.replace(r"[^a-z,A-Z]+", " ", regex=True)
         .str.strip()
         .str.lower()
-        .str.replace(r"\s+", " ", regex=True)
+        .str.replace(pudl.helpers.UNICODE_WHITESPACE_REGEX, " ", regex=True)
         .str.title()
     )